Use Ory Middleware to serve auth from app (#114)
**Related Issue(s):** #

closes #110

**Proposed Changes:**

This PR makes use of the Ory Middleware we have configured on the
application to proxy the auth logic from our own application. This means
that we will no longer need a custom domain for Ory (e.g.
`auth.source.coop`). This simplifies configuration and avoids a handful
of redirect complications that come with the dynamic URLs generated by
Vercel's previews. However, the downside is that anyone's saved
passwords / passkeys will not be directly available as the domain for
auth will have changed.

**PR Checklist:**

- [ ] This PR is made against the dev branch (all proposed changes
except releases should be against dev, not main).
- [ ] This PR has **no** breaking changes.
- [ ] I have updated or added new tests to cover the changes in this PR.
- [ ] I have updated the version in `package.json` to follow the
[Semantic Versioning](https://semver.org/) guidelines.
- [ ] All tests are passing.
- [ ] I have added my changes to the
[CHANGELOG](https://github.com/source-cooperative/source.coop/blob/dev/CHANGELOG.rst)
      **or** a CHANGELOG entry is not required.
- [ ] This PR affects the [Source Cooperative Data
Proxy](https://github.com/source-cooperative/data.source.coop),
      and I have opened issue/PR #XXX to track the change.