Add working Create Organization form (#112)
**Related Issue(s):** #

Closes #113

**Proposed Changes:**

1. This PR adds some foundational work in generating forms with [server
actions](https://nextjs.org/docs/app/getting-started/updating-data) to
better integrate with the NextJS ecosystem. This will reduce the amount
of APIs needed to support the frontend.
2. There were no flags set on an account when they were transferred over
to the new DynamoDB tables. I have updated the script to embrace a set
of defaults for each user. These should be confirmed & the migrations on
production should be rerun once this PR is merged.
https://github.com/source-cooperative/source.coop/blob/c1e5a77b59bb3af416a452908bb1b3947adc784e/src/types/shared.ts#L31-L36