feat: support uploading nested directories (#179)
## What I'm changing

This feature enables a user to upload an entire directory of files,
including files within its nested directories



https://github.com/user-attachments/assets/de867385-8060-45b5-8537-c1af71fd2916



## How I did it

We make use of the
[`webkitdirectory`](https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/webkitdirectory)
flag on the `input` element to allow a user to select a directory
instead of files. Annoyingly, there doesn't seem to be a way to permit
users to select _both_ directories and files. As such, I've created a
separate button for uploading only directories. For these same reasons,
drag-n-drop doesn't support directories (ie drag-n-drop can only support
either directories or files, not both; leaving as just supporting files;
read more
[here](https://github.com/react-dropzone/react-dropzone/discussions/1157#discussioncomment-2537815))

## How you can test it

<!--
  Instructions on how a reviewer can verify these changes.

  Consider including screenshots or video demonstrating feature.
-->