fix: correct opengraph representations (#187)
## What I'm changing

<!--
  High-level summary of what is achieved by these changes.

  Optional: reference related issues.
-->

This PR correctly builds out the [OpenGraph Metadata](https://ogp.me/)
for the product and account pages.

## How I did it

<!--
  Lower-level details of the steps taken to achieve goal.

  Include:
    - considerations made when deciding how to implement feature
    - any changes to API that could impact the Data Proxy
-->

This was more intensive than expected. 

NextJS supports using a filename convention[^opengraph-filename]
(`opengraph-image.tsx`) for opengraph images however I felt it simpler
to expand the already-existing `/api/og` API endpoint[^opengraph-api] to
dynamically generate OpenGraph images. This was largely as it was
awkward to make use of the filename convention with the catch-all route
we use for browsing product files.

## How you can test it

<!--
  Instructions on how a reviewer can verify these changes.

  Consider including screenshots or video demonstrating feature.
-->

Either share a link to yourself on Slack or review the metadata via a
helper tool like opengraph.dev:

* `cholmes` ->
https://opengraph.dev/panel?url=https%3A%2F%2Fsource-cooperative-git-fix-opengraph-radiantearth.vercel.app%2Fcholmes
* `cholmes/overture` ->
https://opengraph.dev/panel?url=https%3A%2F%2Fsource-cooperative-git-fix-opengraph-radiantearth.vercel.app%2Fcholmes%2Foverture

[^opengraph-filename]:
https://nextjs.org/docs/app/api-reference/file-conventions/metadata/opengraph-image#generate-images-using-code-js-ts-tsx
[^opengraph-api]: https://vercel.com/docs/og-image-generation