feat: backfill DOIs (#189)
## What I'm changing

Backfill legacy DOIs and represent them on the UI.


**Example**
<img width="1149" height="448" alt="image"
src="https://github.com/user-attachments/assets/c73e14bf-6d44-4d23-ba27-e744b16734be"
/>


## How I did it


1. Add a script to backfill DOIs directly into DynamoDB. We're storing
the value as an optional `product.metadata.doi` attribute
1. Conditionally display the DOI as mono text within the product
metadata card. The DOI seems to be a bit constricted on the metadata
card, not sure if this is an ideal way to represent this value. It
likely does not make sense to render the value as a clickable link as it
should take the user back to the same product on source.coop


<!--
  Lower-level details of the steps taken to achieve goal.

  Include:
    - considerations made when deciding how to implement feature
    - any changes to API that could impact the Data Proxy
-->

## How you can test it

<!--
  Instructions on how a reviewer can verify these changes.

  Consider including screenshots or video demonstrating feature.
-->


https://source-cooperative-git-feat-backfill-dois-radiantearth.vercel.app/nasa/floods

---

closes #158