Fix membership invite perms (#196)
## What I'm changing

This PR alters the permission lookup to correctly check if a non-admin
user has permissions to edit the memberships of an organization or
product.

#194

<!--
  High-level summary of what is achieved by these changes.

  Optional: reference related issues.
-->

## How I did it

Pairing with @PowerChell, we noticed that we were incorrectly checking
the `Actions.InviteMembership` checks against the `account` rather than
a `membership` instance. Furthermore, it seemed unnecessary to create an
entire mock membership when we only need to check a user account (the
`principal`) against a resource's account id (`membership_account_id`)
and possibly the product id (`repository_id`).

We reworked the typing of `isAuthorized` to dynamically check the
resource based on the action being tested. This adds much better type
safety to the permission checks.

<!--
  Lower-level details of the steps taken to achieve goal.

  Include:
    - considerations made when deciding how to implement feature
    - any changes to API that could impact the Data Proxy
-->

## How you can test it

As a non-admin, navigate to an org or product that you own and attempt
to invite a user to become a member.

<!--
  Instructions on how a reviewer can verify these changes.

  Consider including screenshots or video demonstrating feature.
-->

---------

Co-authored-by: PowerChell <michelle@radiant.earth>